+++ spec/controllers/conversations_controller_spec.rb
@@ -259,7 +259,7 @@ describe ConversationsController, :type => :controller do
       it 'should set response with success to false and message to fail due to no contact' do
         post :create, @hash
         expect(assigns[:response][:success]).to eq(false)
-        expect(assigns[:response][:message]).to eq(I18n.t('conversations.create.no_contact'))
+        expect(assigns[:response][:message]).to eq(I18n.t("javascripts.conversation.create.no_recipient"))
       end
     end
 
@@ -300,12 +300,6 @@ describe ConversationsController, :type => :controller do
       @conversation = Conversation.create(hash)
     end
 
-    it 'succeeds with js' do
-      xhr :get, :show, :id => @conversation.id, :format => :js
-      expect(response).to be_success
-      expect(assigns[:conversation]).to eq(@conversation)
-    end
-
     it 'succeeds with json' do
       get :show, :id => @conversation.id, :format => :json
       expect(response).to be_success
@@ -316,6 +310,28 @@ describe ConversationsController, :type => :controller do
     it 'redirects to index' do
       get :show, :id => @conversation.id
       expect(response).to redirect_to(conversations_path(:conversation_id => @conversation.id))
+    end
+  end
+
+  describe "#raw" do
+    before do
+      hash = {
+        author:              alice.person,
+        participant_ids:     [alice.contacts.first.person.id, alice.person.id],
+        subject:             "not spam",
+        messages_attributes: [{author: alice.person, text: "cool stuff"}]
+      }
+      @conversation = Conversation.create(hash)
+    end
+
+    it "returns html of conversation" do
+      get :raw, conversation_id: @conversation.id
+      expect(response).to render_template(partial: "show", locals: {conversation: @conversation})
+    end
+
+    it "returns 404 when requesting non-existant conversation" do
+      get :raw, conversation_id: -1
+      expect(response).to have_http_status(404)
     end
   end
 end